move some code into a helper class
